Редактирование: {{testName}} fix this
I don't like the raspberry color in some places in the template; it doesn't fit well.

For the test:
I want the correct option in the template to be highlighted in green. A clear notification should be displayed below the test indicating correct/incorrect, without obstructing the view of the answers. There should be more animations.