You didn't quite understand the idea of templates: templates allow for the creation of SETS of different test options in a similar design. Templates can contain various variations - drag-and-drop, multiple-choice, and many other options that the user wants and programs. However, there should be a basic template as an example with very different variations of tasks.
Some notifications do not have translations.

For some reason, Matching (Text-to-Text) and Drag & Drop (Text-on-Text) do not drag and drop properly; all tasks need to be fixed to make them functional, and a basic template with much greater diversity of task types should be added.

In addition to HTML and CSS, JavaScript should be included. Customization should be at a level that allows for the creation of any task with any design. The text for correct and incorrect answers should be defined in JavaScript; nothing from the tests should be in a separate language file, as the user will configure that. Everything related to the test, including all animations and the logic of operation, should be customizable by the user.